1. Implement loading/saving of clumpIds with SpherePack (#189) 2. Handle clumps in SpherePack.fromSimulation (toSimulation raises NotImplementedError with clumps for now) 3. Implement PsdClumpGenerator (#188), clumps are defined with a new ClumpDef class (#185) 4. Require clumps to define memberIds always (otherwise their particles cannot be found easily, which is a problem for exporting to SpherePack and for deleting the particles)

+++ pkg/dem/Funcs.cpp
@@ -266,7 +266,7 @@ vector<shared_ptr<Particle>> DemFuncs::importSTL(const string& filename, const s
 			for(int i=0; i<numFaces; i++){
 				in.read(reinterpret_cast<char*>(&faces[i]),50);
 				const auto& f(faces[i]);
-				LOG_TRACE("binary STL: face #"<<i<<" @ "<<in.tellg()-50<<": normal ("<<f.normal[0]<<", "<<f.normal[1]<<", "<<f.normal[2]<<"), "
+				LOG_TRACE("binary STL: face #"<<i<<" @ "<<(int)in.tellg()-50<<": normal ("<<f.normal[0]<<", "<<f.normal[1]<<", "<<f.normal[2]<<"), "
 					" vertex 0 ("<<f.v[0]<<", "<<f.v[1]<<", "<<f.v[2]<<"), "
 					" vertex 1 ("<<f.v[3]<<", "<<f.v[4]<<", "<<f.v[5]<<"), "
 					" vertex 2 ("<<f.v[6]<<", "<<f.v[7]<<", "<<f.v[8]<<"), "
